+++ janitor/functions/select.py
@@ -1,12 +1,18 @@
 from typing import Any
 import pandas_flavor as pf
 import pandas as pd
-from janitor.utils import deprecated_alias
+from janitor.utils import refactored_function
+from janitor.utils import check, deprecated_alias
 from janitor.functions.utils import _select, DropLabel  # noqa: F401
 
 
 @pf.register_dataframe_method
-@deprecated_alias(search_cols="search_column_names")
+@refactored_function(
+    message=(
+        "This function will be deprecated in a 1.x release. "
+        "Please use `jn.select` instead."
+    )
+)
 def select_columns(
     df: pd.DataFrame,
     *args: Any,
@@ -29,6 +35,11 @@ def select_columns(
         The preferred option when selecting columns or rows in a Pandas DataFrame
         is with `.loc` or `.iloc` methods.
         `select_columns` is primarily for convenience.
+
+    !!!note
+
+        This function will be deprecated in a 1.x release.
+        Please use `jn.select` instead.
 
     Examples:
         >>> import pandas as pd
@@ -209,19 +220,26 @@ def select_columns(
             a callable,
             or variable arguments of all the aforementioned.
             A sequence of booleans is also acceptable.
-            A dictionary can be used for selection on a MultiIndex on different levels.
+            A dictionary can be used for selection
+            on a MultiIndex on different levels.
         invert: Whether or not to invert the selection.
-            This will result in the selection of the complement of the columns
-            provided.
+            This will result in the selection
+            of the complement of the columns provided.
 
     Returns:
         A pandas DataFrame with the specified columns selected.
     """  # noqa: E501
 
-    return _select(df, args=args, invert=invert, axis="columns")
+    return _select(df, columns=list(args), invert=invert)
 
 
 @pf.register_dataframe_method
+@refactored_function(
+    message=(
+        "This function will be deprecated in a 1.x release. "
+        "Please use `jn.select` instead."
+    )
+)
 def select_rows(
     df: pd.DataFrame,
     *args: Any,
@@ -242,12 +260,16 @@ def select_rows(
 
     !!! info "New in version 0.24.0"
 
-
     !!!note
 
         The preferred option when selecting columns or rows in a Pandas DataFrame
         is with `.loc` or `.iloc` methods, as they are generally performant.
         `select_rows` is primarily for convenience.
+
+    !!!note
+
+        This function will be deprecated in a 1.x release.
+        Please use `jn.select` instead.
 
     Examples:
         >>> import pandas as pd
@@ -275,20 +297,27 @@ def select_rows(
             a callable,
             or variable arguments of all the aforementioned.
             A sequence of booleans is also acceptable.
-            A dictionary can be used for selection on a MultiIndex on different levels.
+            A dictionary can be used for selection
+            on a MultiIndex on different levels.
         invert: Whether or not to invert the selection.
-            This will result in the selection of the complement of the rows
-            provided.
+            This will result in the selection
+            of the complement of the rows provided.
 
     Returns:
         A pandas DataFrame with the specified rows selected.
     """  # noqa: E501
-    return _select(df, args=args, invert=invert, axis="index")
+    return _select(df, rows=list(args), invert=invert)
 
 
 @pf.register_dataframe_method
+@deprecated_alias(rows="index")
 def select(
-    df: pd.DataFrame, *, rows: Any = None, columns: Any = None
+    df: pd.DataFrame,
+    *args,
+    index: Any = None,
+    columns: Any = None,
+    axis: str = "columns",
+    invert: bool = False,
 ) -> pd.DataFrame:
     """Method-chainable selection of rows and columns.
 
@@ -302,6 +331,8 @@ def select(
 
     Selection can be inverted with the `DropLabel` class.
 
+    Optional ability to invert selection of index/columns available as well.
+
 
     !!! info "New in version 0.24.0"
 
@@ -311,6 +342,12 @@ def select(
         The preferred option when selecting columns or rows in a Pandas DataFrame
         is with `.loc` or `.iloc` methods, as they are generally performant.
         `select` is primarily for convenience.
+
+    !!! abstract "Version Changed"
+
+        - 0.26.0
+            - Added variable `args`, `invert` and `axis` parameters.
+            - `rows` keyword deprecated in favour of `index`.
 
     Examples:
         >>> import pandas as pd
@@ -323,13 +360,13 @@ def select(
         cobra               1       2
         viper               4       5
         sidewinder          7       8
-        >>> df.select(rows='cobra', columns='shield')
+        >>> df.select(index='cobra', columns='shield')
                shield
         cobra       2
 
         Labels can be dropped with the `DropLabel` class:
 
-        >>> df.select(rows=DropLabel('cobra'))
+        >>> df.select(index=DropLabel('cobra'))
                     max_speed  shield
         viper               4       5
         sidewinder          7       8
@@ -339,23 +376,54 @@ def select(
 
     Args:
         df: A pandas DataFrame.
-        rows: Valid inputs include: an exact label to look for,
+        *args: Valid inputs include: an exact index name to look for,
             a shell-style glob string (e.g. `*_thing_*`),
             a regular expression,
             a callable,
             or variable arguments of all the aforementioned.
             A sequence of booleans is also acceptable.
-            A dictionary can be used for selection on a MultiIndex on different levels.
+            A dictionary can be used for selection
+            on a MultiIndex on different levels.
+        index: Valid inputs include: an exact label to look for,
+            a shell-style glob string (e.g. `*_thing_*`),
+            a regular expression,
+            a callable,
+            or variable arguments of all the aforementioned.
+            A sequence of booleans is also acceptable.
+            A dictionary can be used for selection
+            on a MultiIndex on different levels.
         columns: Valid inputs include: an exact label to look for,
             a shell-style glob string (e.g. `*_thing_*`),
             a regular expression,
             a callable,
             or variable arguments of all the aforementioned.
             A sequence of booleans is also acceptable.
-            A dictionary can be used for selection on a MultiIndex on different levels.
+            A dictionary can be used for selection
+            on a MultiIndex on different levels.
+        invert: Whether or not to invert the selection.
+            This will result in the selection
+            of the complement of the rows/columns provided.
+        axis: Whether the selection should be on the index('index'),
+            or columns('columns').
+            Applicable only for the variable args parameter.
+
+    Raises:
+        ValueError: If args and index/columns are provided.
 
     Returns:
         A pandas DataFrame with the specified rows and/or columns selected.
     """  # noqa: E501
 
-    return _select(df, args=None, rows=rows, columns=columns, axis="both")
+    if args:
+        check("invert", invert, [bool])
+        if (index is not None) or (columns is not None):
+            raise ValueError(
+                "Either provide variable args with the axis parameter, "
+                "or provide arguments to the index and/or columns parameters."
+            )
+        if axis == "index":
+            return _select(df, rows=list(args), columns=columns, invert=invert)
+        if axis == "columns":
+            return _select(df, columns=list(args), rows=index, invert=invert)
+        raise ValueError("axis should be either 'index' or 'columns'.")
+    return _select(df, rows=index, columns=columns, invert=invert)