+++ tests/timeseries/test_sort_timestamps_monotonically.py
@@ -47,8 +47,8 @@ def test_sort_timestamps_monotonically_decreasing(timeseries_dataframe):
 def test_sort_timestamps_monotonically_strict(timeseries_dataframe):
     """Test sort_timestamps_monotonically for index duplication handling"""
     df = timeseries_dataframe.shuffle(reset_index=False)
-    random_number = randint(1, len(timeseries_dataframe))
-    df = df.append(
-        df.loc[df.index[random_number], :]
+    random_number = df.index[randint(1, len(timeseries_dataframe))]
+    df = pd.concat(
+        [df, df.loc[[random_number], :]]
     ).sort_timestamps_monotonically(direction="increasing", strict=True)
     assert df.equals(timeseries_dataframe)